+++ CodexLib/Components/FeatureForPrerequisite.cs
@@ -1,0 +1,16 @@
+﻿using Kingmaker.Blueprints;
+using Kingmaker.Blueprints.Facts;
+using Kingmaker.UnitLogic;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace CodexLib
+{
+    public class FeatureForPrerequisite : UnitFactComponentDelegate
+    {
+        public BlueprintUnitFactReference FakeFact;
+    }
+}